# openapi.yaml
openapi: 3.1.0
info:
  title: Powerlifting App API
  version: 1.0.0
  description: |
    REST API for the Powerlifting App — a multi-user workout tracking platform
    for athletes and coaches.

    ## Auth
    All endpoints except `/auth/register`, `/auth/login`, and `/auth/token`
    require `Authorization: Bearer <access_token>`.

    ## Errors
    All error responses use **RFC 7807 Problem Details** with
    `Content-Type: application/problem+json`.

    ## REQ mapping
    Each operation carries an `x-req` extension listing the requirement IDs it
    fulfils (e.g. `x-req: [REQ-001]`).
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# =============================================================================
# COMPONENTS
# =============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    ISODateTime:
      type: string
      format: date-time
      example: "2026-03-11T14:00:00Z"
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: "https://api.example.com/problems/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "email must be a valid email address"
        instance:
          type: string
          format: uri
          example: "/api/v1/auth/register"

    # --- Auth (REQ-001, REQ-002) ----------------------------------------------
...
    UserSummary:
      type: object
      required: [id, email, role, profile]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        email:
          type: string
          format: email
        role:
          type: string
          enum: [athlete, coach]
        profile:
          oneOf:
            - $ref: '#/components/schemas/AthleteProfile'
            - $ref: '#/components/schemas/CoachProfile'
          discriminator:
            propertyName: role
    AthleteProfile:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          example: "Jane Lifter"
        weightClass:
          type: string
          nullable: true
          example: "72kg"
        dateOfBirth:
          type: string
          format: date
          nullable: true
          example: "1995-06-15"
    CoachProfile:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          example: "Coach Mike"
        bio:
          type: string
          nullable: true
          example: "USAPL certified coach, 10 years experience."
...
    Relationship:
      type: object
      required: [id, coachId, athleteId, status, invitedAt]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        coachId:
          $ref: '#/components/schemas/UUID'
        athleteId:
          $ref: '#/components/schemas/UUID'
        status:
          type: string
          enum: [pending, active, declined, removed]
        invitedAt:
          $ref: '#/components/schemas/ISODateTime'
        resolvedAt:
          $ref: '#/components/schemas/ISODateTime'
          nullable: true
    AthleteRosterEntry:
      type: object
      required: [relationshipId, athlete]
      properties:
        relationshipId:
          $ref: '#/components/schemas/UUID'
        athlete:
          $ref: '#/components/schemas/UserSummary'
    InviteAthleteRequest:
      type: object
      required: [athleteEmail]
      properties:
        athleteEmail:
          type: string
          format: email
    RespondInvitationRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, decline]

    # --- Exercises (REQ-006, REQ-007) -----------------------------------------
...
    Program:
      type: object
      required: [id, coachId, name, durationWeeks, daysPerWeek, status, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        coachId:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "12-Week Peaking Block"
        description:
          type: string
          nullable: true
        durationWeeks:
          type: integer
          minimum: 1
          example: 12
        daysPerWeek:
          type: integer
          minimum: 1
          maximum: 7
          example: 4
        status:
          type: string
          enum: [draft, published]
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
        updatedAt:
          $ref: '#/components/schemas/ISODateTime'
    ProgramList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Program'
    CreateProgramRequest:
      type: object
      required: [name, durationWeeks, daysPerWeek]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        durationWeeks:
          type: integer
          minimum: 1
        daysPerWeek:
          type: integer
          minimum: 1
          maximum: 7
    UpdateProgramRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        durationWeeks:
          type: integer
          minimum: 1
        daysPerWeek:
          type: integer
          minimum: 1
          maximum: 7
        status:
          type: string
          enum: [draft, published]

    # --- Program Days (REQ-009) -----------------------------------------------
...
  responses:
    400BadRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/validation-error"
            title: "Validation Error"
            status: 400
            detail: "password must be at least 8 characters"
    401Unauthorized:
      description: Missing or invalid access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Bearer token is missing or expired"
    403Forbidden:
      description: Authenticated but insufficient role or resource ownership.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/forbidden"
            title: "Forbidden"
            status: 403
            detail: "You do not have access to this resource"
    404NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/not-found"
            title: "Not Found"
            status: 404
    409Conflict:
      description: Conflict with existing state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/conflict"
            title: "Conflict"
            status: 409
            detail: "An account with this email already exists"
    422UnprocessableEntity:
      description: Business rule violation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/unprocessable"
            title: "Unprocessable Entity"
            status: 422
            detail: "Sets cannot be edited after the session is closed"

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /coaches/{coachId}/athletes:
    get:
      tags: [Relationships]
      operationId: listCoachAthletes
      summary: List athletes on a coach's roster
      x-req: [REQ-005]
      description: |
        **REQ-005** — Returns all athletes with `status: active` for the given
        coach. Caller must be the coach identified by `coachId`.
      parameters:
        - name: coachId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Active athlete roster.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AthleteRosterEntry'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'

    post:
      tags: [Relationships]
      operationId: inviteAthlete
      summary: Send a coach invitation to an athlete by email
      x-req: [REQ-005]
      description: |
        **REQ-005** — Coach sends an invitation to the athlete with the given
        email address. Creates a relationship with `status: pending`. Returns
        404 if no athlete account is found for that email, 409 if an active or
        pending relationship already exists.
      parameters:
        - name: coachId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteAthleteRequest'
      responses:
        "201":
          description: Invitation sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Relationship'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'
        "409":
          $ref: '#/components/responses/409Conflict'
  /coaches/{coachId}/athletes/{athleteId}:
    delete:
      tags: [Relationships]
      operationId: removeAthleteFromRoster
      summary: Remove an athlete from the coach's roster
      x-req: [REQ-005]
      description: |
        **REQ-005** — Sets the relationship `status` to `removed`. The coach
        can no longer access the athlete's data.
      parameters:
        - name: coachId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: athleteId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Athlete removed from roster.
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'
  /users/{userId}/invitations:
    get:
      tags: [Relationships]
      operationId: listInvitations
      summary: List pending coach invitations for an athlete
      x-req: [REQ-005]
      description: |
        **REQ-005** — Returns all `status: pending` relationships for the given
        athlete. Caller must be the athlete identified by `userId`.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Pending invitations.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Relationship'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
  /invitations/{invitationId}:
    patch:
      tags: [Relationships]
      operationId: respondToInvitation
      summary: Accept or decline a coach invitation
      x-req: [REQ-005]
      description: |
        **REQ-005** — Athlete accepts or declines a pending invitation.
        - `accept` → `status: active`
        - `decline` → `status: declined`

        Returns 422 if the invitation is not in `pending` state.
      parameters:
        - name: invitationId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondInvitationRequest'
      responses:
        "200":
          description: Invitation updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Relationship'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'
        "422":
          $ref: '#/components/responses/422UnprocessableEntity'
  /users/{userId}/coaches:
    get:
      tags: [Relationships]
      operationId: listAthleteCoaches
      summary: List coaches linked to an athlete
      x-req: [REQ-005]
      description: |
        **REQ-005** — Returns all active coach relationships for the athlete.
        Caller must be the athlete identified by `userId`.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Linked coaches.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Relationship'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'

  # ---------------------------------------------------------------------------
  # EXERCISES  ·  REQ-006, REQ-007
  # ---------------------------------------------------------------------------
...
  /programs:
    get:
      tags: [Programs]
      operationId: listPrograms
      summary: List programs authored by the caller
      x-req: [REQ-008]
      description: |
        **REQ-008** — Returns all programs owned by the authenticated coach.
        Coach role required.
      responses:
        "200":
          description: Program list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramList'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'

    post:
      tags: [Programs]
      operationId: createProgram
      summary: Create a new program (saved as draft)
      x-req: [REQ-008]
      description: |
        **REQ-008** — Creates a program with `status: draft`. Coach role required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProgramRequest'
      responses:
        "201":
          description: Program created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
  /programs/{programId}:
    get:
      tags: [Programs]
      operationId: getProgram
      summary: Get a program by ID
      x-req: [REQ-008, REQ-011]
      description: |
        **REQ-008, REQ-011** — Coach can retrieve their own program; athlete
        can retrieve a program they are assigned to.
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Program detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'

    patch:
      tags: [Programs]
      operationId: updateProgram
      summary: Update a program
      x-req: [REQ-008]
      description: |
        **REQ-008** — Coach can edit a program that has no active assignments.
        Returns 422 if the program has active assignments.
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProgramRequest'
      responses:
        "200":
          description: Program updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'
        "422":
          $ref: '#/components/responses/422UnprocessableEntity'

    delete:
      tags: [Programs]
      operationId: deleteProgram
      summary: Delete a program
      x-req: [REQ-008]
      description: |
        **REQ-008** — Coach can delete a draft program with no active
        assignments. Returns 422 if active assignments exist.
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Program deleted.
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'
        "422":
          $ref: '#/components/responses/422UnprocessableEntity'

  # ---------------------------------------------------------------------------
  # PROGRAM DAYS  ·  REQ-009, REQ-011
  # ---------------------------------------------------------------------------
...